Admin menu: introduce a function to get the main menu item's slug
We switch the main menu item from the configuration page to the modules
page when the plugin is network active in multisite (see #1). Formerly,
we were keeping the slug of the main menu item the same, essentially
just switching the text of the menu. That caused problems, because the
module’s page basically ended up having two slugs. Getting it to show
up as the current menu item for both wasn’t really possible. In short,
we would have had to make all of the links to it dynamic, or set the
current menu item with JS. Both of those solutions seemed much more
“hacky” and complex than necessary. Now we just have a function to call
which will return the slug of the main menu item, whether it is the
configuration or modules screen.