Multiste: store user points separate for different sites when not network-active
When network active, as on non-multisite, the points for a user are
stored in user meta under a key universal to the whole network. Now
when the plugin is active only on single sites on a network, the points
for the user are stored separately for each site, by prefixing the meta
key with the blog prefix.

See #1