Fix a randomly failing unit test
This unit test made the wrong assumption that the default order was
`ASC`, when it is really `DESC`. It was randomly failing (though
passing most of the time), because it was using the default `orderly`
which is `date`. The two transactions are added so close together that
the `date` order is unreliable. Change it to `id`, which will always
work.

Fixes #80