More tweaks to the Travis CI config
Run PHP 5.3 first, so that the `codesniff` pass will run first. That
way we fail ASAP.

Run the tests for 3.7 Multisite as well.

Don’t run the 3.7 tests on PHP 5.5, it will cause errors because
`$wpdb` doesn’t support `mysqli` until 3.8, and `mysql` is deprecated
in PHP 5.5.

Simplify the tests config for PHP 5.6.

Se #94