Better docs for hook, and don't validate constant
There’s no easy way to just check if a constant’s value is falsey
without using fuzzy equality checking. I don’t think there is really a
need to make this check anyway. If a user defines this constant, it
aught to have a sane value. The onus is on them.

Also adds better docs to the filter.

See #92