Set the charset when creating the database
Includes the following database schema changes:

- We now set the charset and collation, like WordPress core does.
- We don’t try to specify a name for `PRIMARY` keys, since they can’t
be named and `dbDelta()` chokes on them.
- The `points_type` and `log_type` fields of the points logs table are
`VARCHAR(255)` instead of `256`, for compatibility with MySQL 5.0.3,
and because `dbDelta()` tries to create duplicate keys otherwise (the
key names are returned from the database with the length, 255, appended
to them, and don’t match the names in the SQL schema).

The SQL schema has been moved into a function, and is now fully
parseable by `dbDelta()`, which means that it can be used to update the
database, not just create it.

See #97

(cherry picked from commit 98595862ae1bc49d8577e3f9fb0f95bc3c7e10d4)