Change rank_id column name in rankmeta table
WordPress assumes that the object ID column in a meta table with be of
the form `{object_type}_id`. Because our object type is
`wordpoints_rank` in this case, the column name needs to be
`wordpoints_rank_id` instead of just `rank_id`.

It is possible to implement a hacky workaround instead, (which we do
for the points logs meta table,) but I’ve opted for this solution since
it is simpler.

See #122