Add initial ranks API
This will receive more iteration as the UI is built. Here is how the
API works:

The basic elements of the ranks component are of course the ranks.
These objects are stored in the ranks database table and are
represented within the code using the `WordPoints_Rank` class, which
parallels in many ways `WP_Post`.

Ranks may have metadata in addition to their basic fields. This is
stored in the rank meta table, and may be
added/updated/deleted/retrieved with procedural functions.

Ranks are categorized into different ‘types’. Each rank type has its
own extra metadata and rules for when a user can achieve that rank.
Each rank type is handled by a child of the `WordPoints_Rank_Type`
class. This will handle the display of the form fields for ranks of
this type to the user, as well as validating the ranks metadata before
they are saved/updated. It will also hook into actions to trigger
possible rank transitions for a user, and determines whether a user can
move up to a rank or not.

The rank types are registered with and contained in the
`WordPoints_Rank_Types` class, in much the same fashion that points
hook types are registered with the `WordPoints_Points_Hooks` class.

Ranks are also organized into ‘groups’. These are sets of ranks
arranged in an ascending order, like a ladder the user may climb until
they reach the top. Rank groups are represented by instances of the
`WordPoints_Rank_Group` class, and are registered with the
`WordPoints_Rank_Groups` class. Each rank group can contain ranks of
various types, but for a type to be used with a group it must be
registered with it using
`WordPoints_Rank_Groups::register_type_for_group()`.

See #122