Make rank type display meta $meta default array()
The `display_rank_meta_form_fields()` method of the
`WordPoints_Rank_Type` class accepts a single argument, `$meta`. This
makes it so that the argument defaults to an empty array. This is done
so that it is more obvious to developers that they can’t expect the
metadata to be set when the function is called.

See #122