Ranks screen JS: fix bad line break JSHint errors
These warnings are because automatic semicolon insertion (ASI) can
break things when semicolons get inserted in the wrong place. That can
happen, for example, when a line ends in a statement or variable, and
the rest of the code, for example `+`, is on the next line. Really, ASI
should be smart enough to know that semicolons never need to be
inserted within parenthetical statements. However, JS tools generally
don’t account for parenthesis correctly. This is a shame, because it
means you have to put the `+` or `&&` operator at the end of the line,
instead of after the line break. That makes wrapped code more difficult
to read, because you have to scroll to the end of the line to see what
operator was used. In conclusion, let’s just not wrap long things the
way that we might in PHP. At least for now.