Refresh user ranks when moving a rank in a group
1. When a rank is added to a group, we check if any of the users in the
rank in the position below it should be moved up to it.
2. When a rank is moved within the group, we first move users to the
rank directly below it. Then, after moving it, we check if any of the
users of the rank that is now below it should be moved up to it.
3. When a rank is removed from the group, we check if there is a rank
below it. If so, we assign users with the rank being removed to that
rank. If not, we delete those rows completely.

See #155