Update user rank func: improved base rank handling
When updating a user’s rank, we have to account for the fact that when
users have a base rank, they currently don’t always have an entry in
the database. Because of that, we can’t rely on an update query to
modify the user’s rank; sometimes we need to run an insert instead.

We were doing this, but we were making the assumption that if the user
had a base rank would always need to do an insert. However, now
sometimes the user’s rank will be in the table, so we need to fall back
to the usual update query.

See #155