Recheck if users can hold a rank when updating it
When we update a rank, it is possible that users who used to hold the
rank cannot any longer. When the update means that the rank’s group or
group position is being changed, this is automatically handled by the
group classes (f3e0d02a14c010d4d87ac9a444f43555716b38b4). But, in order
for it to take into account the updated metadata, it needs to happen
after the metadata is added. So we switch the order there.

Also, the metadata can be changed even when the rank position isn’t.
This needs to be handled separately, using
`wordpoints_refresh_rank_users()`.

Includes unit tests.

See #155