Pass points logs through wp_kses()
Closes off any potential XSS vulnerabilities.

This does not fix any currently known XSS vulnerabilities, other than
one which could only be exploited by users who have the Administrator
or Editor roles. [That isn’t considered to be a vulnerability at all on
the front
end](https://make.wordpress.org/core/2010/12/31/the-published-exploit-fo
r-wordpress-3-0-4-isnt-accurate/), because those roles have the
`unfiltered_html` capability. However, because the logs are also
displayed in the admin area, it could present a minor threat.