Don't try to skip uninstalling points component
We were previously attempting to detect if it was installed, and only
run the uninstaller if so. That’s not reliable, because of the use of
WordPoints “network” option feature. This selects a regular or site
option based on how WordPoints is activated, but it isn’t possible to
determine that on uninstall, since the plugin isn’t active at all.

See #68