Fix points hook multi-number assignment
When new points hooks are created they need to be assigned a unique ID,
which is comprised of a base slug for the hook type and an ascending ID
number. The ID number for a hook template needs to be `__i__`, which
will be replaced with the incremented multi number when a new hook is
created from the template.

During cleanup in #77, we’ve changed the way the template ID in a hook
template was generated. The result was that the `$number` was being
used instead of `__i__`. The number would in the case of a hook
template be 0.

The result of this inadvertent change is that the hook would be saved,
but it would not fire, because the 0 hook number is reserved for hook
templates.

This fixes the problem by setting the hook ID in templates
appropriately without respect to the value of the `$number`.
Previously, we only attempted to correct this if the `$number` was set
to a non-empty value, which meant that these things had to be handled
each time a hook template was displayed.

See #77