Introduce a comment received points hook
It will award a post author points when they receive a comment on one
of their posts. If the comment it removed, the transaction will be
automatically reversed.

To keep DRY, we introduce a new base class for comment approved points
hooks, and also use it for the Comment points hook.

See #111