Travis CI: change PHPUnit order
On PHP 5.2 we need to remove the exclude groups `ajax` and `uninstall`
from the XML config file because of a bug in the legacy version of
PHPUnit that runs on that PHP version. We need to run the general tests
for multisite before doing that, otherwise those groups, which are
normally excluded, will be run as well. We run those groups each on
their own after running the general tests.

See #143