Travis: get latest uninstall tester on PHP 5.2
We can’t use composer for dependencies on PHP 5.2, since it requires
5.3. Instead, we have to download the uninstall tester manually.
However, we had hardcoded the version number in the URL. Now we’re
using git to checkout the latest version. This doesn’t guarantee that
it will be the same version specified in composer.json, but at least we
won’t need to update it constantly.

See #143