Fix points logs cleaning tests
These tests were checking that log meta was being deleted, but because
they were using `WordPoints_Points_Logs_Query`s they weren’t actually
testing whether the metadata had been deleted, because the query will
only work if the log still exists, and in this case, it doesn’t. What
we need to do is get the IDs of the logs and then retrieve all of their
metadata with the get meta function.

See #226