Extend base widget class for the points widgets
The base points widget class now extends the new `WordPoints_Widget`
class. There are two significant things here:

1. The class is intended to be a parent only, but wasn’t declared
abstract. This is strange, and I’ve added a note about it in the class
docs. Making it abstract is technically a breaking change, so maybe
we’ll wait for 2.0.0 to do that.
2. I’ve implemented the `verify_settings()` function, which will
automatically check the validity of a points type setting if one is
expected.

Another oddity is that there was a redundant constructor in the class,
that really did nothing. That is no more.

See #230, #71