PHPUnit tests for the My Points widget
The widget now uses the new bootstrapped API, to be DRY. This also
results in the user receiving an error message if an invalid points
type is being used (see #230).

In addition, a bug is fixed with the default text generation. If the
text was completely empty we generated some text for the widget using
the points type name. However, the arguments being passed to
`wordpoints_get_points_type_setting()` were backward, which means that
nothing was being returned. The result was something like “: $40”. I’m
actually unsure whether we should keep this behavior at all. If the
user wants to keep the widget empty, that’s their prerogative. But that
would be for another ticket.

See #71