Add a function to verify an nonce
WordPress has a function to verify an nonce, however, it is pluggable
and so can’t be relied upon to perform proper sanitization. There is no
real precedent for the best way to sanitize an nonce, so we’ve
introduced this function so that all nonce sanitization will be done in
just one place. That way, if it needs to be modified in the future,
we’ll be all set.

Includes unit tests.

For PHPCS, and early sanitization, with love.

See #231 as an excuse.