Logs queries: only retrieve the current page
Before are cache was designed in such a way that when it was primed
when a query was constructed the entire result set was retrieved from
the database and saved to the object cache. This meant that we were
pulling results from the database that we might not ever use,
especially when the results were being paginated.

This changes the `prime_cache()` method so that it just sets up the
query to use cacheing—it doesn’t actually pull the results into the
cache. That doesn’t happen until `get()`, `count()` or `get_page()` are
actually called later on. That way we’re only caching the results that
we would actually be pulling form the database anyway.

The result of this is that each page is cached separately. We could
have gone the complete route of trying to save the paginated results to
the cache, but when you have different caches using different numbers
of results per page, that can quickly get overly complex. So yes, the
same query will use two different caches for results paginated
differently. However, if the entire query has been cached, that main
cache will be used and the results won’t be cached separately for
pagination at all.

We might revisit this in future, but at the moment it is an
improvement, and for the most part won’t result in duplicated cache
records.

Fixes #113