Only allow disabling auto-reversal on legacy sites
Only on sites where disabling auto-reversal of the Comment hook is
necessary because of existing hooks do we allow it to be disabled as a
legacy feature.

We also introduce a new method to the base hook class, `set_option()`,
and include unit tests for it and related methods.

See #221