PHPUnit: update tests for Comment Removed hook
The tests involving the Comment and Comment Removed hooks need to be
updated now that the Comment Removed hook is disabled by default.

We now need to manually register the Comment Removed hook in any test
that requires it. However, the only way to do that is by calling
`WordPoints_Points_Hooks::initialize_hooks()` again after calling the
`register()` method to register the hook’s class. This results in a
second set of hook objects being instantiated. This effectively causes
the hooks to be fired twice in the tests that do this. This state
doesn’t leak into other tests, because the actions are reset after each
test.

It does cause an issue with the Comment hook in regard to
auto-reversal. The auto-reversal feature can be disabled on legacy
installs by setting a certain hook type “option”. This is object
instance-specific, and is saved in a property. This causes problems
when we need to reinitialize the hooks, because then the objects
returned by the `get_handler_by_id_base()` method are different from
the objects which are hooked to the actions. This state of things will
continue to affect each test after a re-initialization occurs.

Because of this, I’ve found it necessary to isolate the Comment
Removed-related tests that re-initialize the hooks, and add them to a
new excluded `legacy` test group.

We don’t currently run this group on Travis, and I’m going to leave it
like that for now.

See #221