Make sure comment_disapprove logs are generated
Now that the Comment Removed hook may be disabled on legacy installs,
we need to provide a separate function to generate the log text, in
case the logs get regenerated for some reason.

This adds a new function to do that, and it is now used by the Comment
Removed hook class’s `logs()` method.

Includes unit tests.

See #221