Post hooks: move logs cleaning to base class
We were duplicating this method in both the Post points hook class and
the Comment Approved base class. By moving it up from the Post class,
we not only DRY ourselves out a bit, but we also add a new feature to
the Comment Approved hooks: when the post is deleted the post type is
saved. The post hook already did this.

See #256