Post points hooks: Add a bootstrap for reversals
We add a method to retrieve reversals and another to perform the
reversal. This reduces WETness. We now use this bootstrap in the Post
and comment approved hooks.

We also fix a bug that was causing logs for comments to be retrieved
from the database unnecessarily.

Includes unit tests.

Fixes #252