Display rank field values when placeholders off
This is the way it was supposed to work, but the condition was
convoluted, and ultimately incorrect. This is simpler and actually
works.

The reason that this wasn’t actually causing issues anywhere is that we
always use placeholders at the moment (though they are off by default).

Working on #269