Update un/installer after removing back-compat
After removing some backward-compatibility code in the base
un/installer class and the components class, the uninstallation tests
were giving errors. The problem was two-fold: first, the component’s
weren’t being registered, because the `do_action()` call had
accidentally been marked as just back-compat code, and so was removed.
Secondly, once this was fixed, it was revealed that the points
component wasn’t loading all of the dependencies needed during
uninstall. This had changed because we’ve stopped loading the points
functions in the same file as the general functions of the points
component.

See #281