Promote users to the correct rank
This simplifies the rank transitioning code, and fixes a bug where
every other rank would be skipped when a user was being promoted. This
would cause a problem when you had ranks like this:

0 points (base)
100 points
1000 points

If a points-less user was awarded, say 150 points, they would be
transitioned to the 1000 points rank. Here is how it would happen. The
maybe-increase-rank function would first check if the user could
transition to the 100 points rank. It would get the affirmative. That
is the correct behavior. Then it would proceed to check if the user
could transition to the next rank. However, it would actually check one
rank above that. Of course, in a case like this, there is no fourth
rank, so the third rank for 1000 points would be returned.

Adds unit tests.

Fixes #300