Add context property to base un/installer class
We set this property to the context of the current operation
(‘network’, ‘site’, or ‘single’), whether that operation is
un/installation or an update. We’ll use this in some of our methods as
we provide a better bootstrap for uninstallation. Specifically, we’ll
be introducing an `uninstall_metadata()` method, which will need to
know the context so it can automatically prefix the meta key with the
blog prefix when the context is ‘site’. We’ll then be using this method
as part of our work on deleting screen option metadata during
uninstallation.

See #306, #296