Un/installer: be less abstract
Previously a bunch of the methods were abstract. Specifically;
`(un)?install_(single|site|network)`. After the recent work on
providing a better un/install bootstrap, the idea is that now these
functions won’t need to be implemented in most cases. So keeping them
as abstract would be silly. The only one that was still abstract was
`install_network()`.

This makes `install_network()` no longer abstract and updates the
inline docs.

I’ve also dropped the `abstract` from the `load_dependencies()`
function. I was the final abstract function in the class, but there is
no need for it to be, since it’s probable that dependencies will be
unneeded for at least some projects.

See #306, #307