Fix errors thrown by the WPCS CSRF sniff
None of these locations are vulnerable. Two have nonce checks later on,
in the file included within the statement. The other is hooked to an
action that is called only after nonce validation on the settings
screen. However, I opted to add an explicit nonce check here, for good
measure.

See #324