Un/install: automatically set db version
When a module/component is installed, we now automatically set the
version in the database as part of our provided bootstrap. In addition,
other parts of the bootstrap are simplified and improved, and many
tests are added. The main other change is that the `$option_prefix`
property is now deprecated in favor of new `$type` and `$slug`
properties.

Two new bits of candy got added to our testing tools along the way:

1. When a test requires multisite, you can now just add `@requires
WordPress multisite` in the method/class docblock, and it will
automatically be skipped if multisite is not enabled. This can also be
inverted with `@requires WordPress !multisite`. You can do the same
thing when you need WordPoints to be/not to be network-active:
`@requires WordPoints network-active`.
2. The `WordPoints_Mock_Filter` class was added, to help with testing
filters (more will be added here in future).

See #307