Uninstall modules & components before core
We need to uninstall modules and components before uninstalling
WordPoints itself, rather than as part of the uninstallation process.
This avoids the possibility of unexpected results when
components/modules are uninstalled per-site *after* WordPoints is
uninstalled per-site (which was the previous way).

This is exactly what was happening with our new changes to
automatically remove version data from the database. The data was being
removed from each site during WordPoints’s per-site uninstall, but then
being re-added when the components were uninstalled with the network
uninstallation (because the option gets saved back to the database
after the value is unset from the array).

See #307