Fix `@requries` not working for one test
After switching to the new `@requires` annotations, one test was
failing. It was using `@requires WordPoints network-active`, but it was
still running on multisite when WordPoints wasn’t network active, and
thus was failing. This was happening in all PHP versions except 5.2,
which uses an older version of PHPUnit.

Investigation led to the discovery that PHPUnit used to call
`checkRequirements()` after `setUp()`, but now calls it before
`setUp()` instead. It turns out that the test before the failing one
was playing with the database option that determines whether WordPoints
is network-active or not. While the option is deleted when the
`ROLLBACK` query is called in `tearDown()`, the option value was still
in the cache, which isn’t flushed until `setUp()`. Because `setUp()` is
no longer called until after the requirements are checked,
`is_wordpoints_network_active()` was returning true based on the cached
value in newer PHP versions.

This is fixed by manually clearing the cache at the end of that test.