Add function to flush points logs caches
Previously we had a function to clean the cache when a user’s points
were altered, but it was designed specifically to be attached to the
`wordpoints_points_altered` action. This introduces a more generic
function based on the same logic. Now we will be able to use it in
various other places as well.

See #227