Mark all queries that don't need caching
WPCS has a sniff that will flag uncached queries, and this helps us to
see where we are missing caching. We have previously disabled this
sniff, because it wasn’t configurable. I have now fixed that, so we
will be enabling it again.

However, there are some queries which do not need to be cached. For
example, during an update, or when it is a DB schema change. By
whitelisting these queries, we reduce noise, and now we can focus on
the queries that really do need caching.

See #227