Update query args to match new syntax
I happened upon this while working on #227. The `default` points logs
query is being registered with the `cache_queries` arg value
`’results’`. This is the old syntax, which required each type of query
that should be cached to be specified. However, the new syntax is just
to turn caching on or off as a whole, so the `cache_queries` arg is now
a boolean flag. This truth-y value works, but I’ve changed it to `true`
to reduce confusion.