Mark this query as not needing caching
This query is a part of the upgrade to 1.2.0. It deletes any points
logs for deleted users. Normally we’d want to accompany that with a
cache delete call. But caching for the points logs wasn’t introduced
until much later, in version 1.5.0. So there would be no cache to get
deleted.

See #227