Flush points logs caches when a user is deleted
We have a function the removed their old points logs, but it didn’t
flush the caches afterward. It is not important for the cache of that
specific user’s logs to be flushed. The user is being deleted, so that
query wouldn’t be called anymore. But it helps to free up cache space
if we do. Also, we need to flush the other caches, or else the logs
will continue to include the logs for that user.

I’ve updated the unit tests to test this. I’ve also split the multisite
tests for this function from the non-multisite tests. The test method
was too long.

See #227