Flush points logs caches on blog delete
Previously the logs were deleted, but the caches were not flushed
afterward. Now they are. This is not so important for the queries for
that blog, except that they are taking up space. The stale cache won’t
matter because the queries won’t be run anymore. However, it is
important that the cache for the `network` queries is cleared.

See #227