Fix incorrect sanitization of points hooks lists
These lists are an array of comma-delimited hook IDs. We were
sanitizing them with `sanitize_key()`. That is causing the function not
to work correctly, and our tests are failing. The problem is that
`sanitize_key()` will strip out the commas. We need to use
`sanitize_text_field()` instead.

In addition, I’ve refactored the logic to get the list of points types
from the database. Previously, the function was not verifying whether
the points types were valid. Now it will only save the lists for valid
points types.

See #324