Add `@property` and `@method` annotations
This mock for the base un/installer class adds magic getter and setter
methods to allow each of the protected properties to be accessed by the
tests. We add `@property` annotations to the docblock to indicate this.
It also adds a magic caller to allow all of the methods to be called by
the tests, so we add `@method` annotations for each of them.

Working on #330, and had a use for `@method` there, which made me
realize that we could use these annotations here as well.