PHPUnit: Add Ajax callbacks to hook backup
The hooks are backed up the first time the `setUp()` is run, and
restored each time in `tearDown()`. This causes the Ajax callbacks not
to remain hooked to the correct hooks, because they aren’t hooked up
yet the first time `setUp()` is called, due to the file having to be
included manually. This problem did not present itself before, because
there was only one ajax file having to be included, and because it was
loaded in `setUpBeforeClass()`, and was the first test case to run, the
hooks would be backed up. Now we have introduced a second test case,
and so the hooks for the one that runs second won’t be backed up
anymore.

See #330