Update admin notice dismiss to use Ajax
WordPress 4.2 introduced the ability to mark an admin notice as
dismissible by adding the `is-dismissible` class to the notice element.
This adds support for dismissible notices to our template function. It
also updates the code for dismissible notices that are tied to an
option. The option must be deleted for the notice to be dismissed.
Previously, this was done by the user clicking a button to submit a
form. We still do that for non-js users, but when JS is present, we
hide the form and trigger an Ajax action to delete the option when the
user clicks the dismiss button added by WordPress instead. This brings
us better parity with the notice dismissal UI that is familiar to users.

See #222