Remove unnecessary use of md5()
There is no need to use a hash here. We only do so because the code was
inherited from the plugins list table. The code was originally
introduced in [WP@
16345](https://core.trac.wordpress.org/changeset/16345). See
[class-wp-plugins-list-table.16235.diff](https://core.trac.wordpress.org
/attachment/ticket/15318/class-wp-plugins-list-table.16235.diff) and
[comment 6](https://core.trac.wordpress.org/ticket/15318#comment:6) on
the ticket.

We have had trouble before with using the module name to generate the
ID here. Module names aren’t guaranteed to be unique, and so duplicate
element IDs can result at times. This leads to invalid HTML, see #105.
I’ve switched to using the module file, because it will always be
unique. I pass it through `sanitize_key()` to remove any characters
that might cause problems.

In conclusion, the MD5 hashing algorithm is not really
cryptographically safe. This case has nothing to do with cryptography
or security, but by using MD5 here, we may have been inadvertently
encouraging others to use it in unsafe situations.