SECURITY: Don't use MD5 to hash cache keys
The MD5 hashing algorithm is becoming obsolete, and should no longer be
considered cryptographically secure. It is not recommended to use it in
new applications. That said, WordPress still uses it in the `wp_hash()`
function, which is used to generate nonces, among other things.
However, `wp_hash()` incorporates a secret salt into the hash, while in
this case we are using unsalted hashes.

These hashes are used as keys for the points logs query cache. The
values hashed are the SQL query strings. By using the MD5 hashing
algorithm, it might be possible for an attacker to exploit a hash
collision to trick the code into returning the cached results for a
different query. It would be possible to exploit this through queries
that incorporate user-supplied data, because this would allow the
attacker to manipulate the query string to produce a colliding hash.

The result of exploitation would be that the attacker would be able to
return results for a different query. This could result in information
disclosure, but would usually be no more than a minor security concern,
if indeed it would be any concern at all.

Note that this would not allow an attacker to view logs which would
normally be hidden from him by per-log capability checks (for example,
logs for private posts). However, on multisite, it could potentially
allow an attacker to view logs from another site to which he would not
normally have access.

Except in special cases, in which a site admin has not set up proper
per-log capability checks for logs which include sensitive information,
there would be very little incentive for exploitation. Nonetheless, it
is best for us not to use MD5, even if only to set a good example.

The use of PHP’s built in `md5()` function has therefore been replaced
with `wordpoints_hash()`, which currently uses the SHA256 hashing
algorithm.