Don't run module compat. check if install fails
If for some reason WordPoints fails to install itself properly when it
is activated, the version may not be set in the database. This would
cause the breaking-update module compatibility check to be run on every
page load. By checking that the version is set, and bailing if it is
not, we avoid this.

See #349