Un/installer: set the version on reactivation
Previously, the version would only be added during the initial install;
reactivating the plugin would not fix a corrupted version. Now it will.

This is important, because I’ve discovered that we were not installing
the plugin properly in the latest version, leading to some installs not
actually running the core install script and so therefore not having
the version set.

Includes unit tests.

See #349