Install WordPoints when not installed
When we check to see if any plugins/modules need to be updated, we skip
over any that aren’t installed. This is necessary for modules, because
usually that is an indicator that the module doesn’t actually have an
install script.

However, because of a recent bug in WordPoints, sometimes WordPoints’s
own install script will not have executed. Because of this, we should
execute it manually when we detect that it hasn’t been done.

Includes unit tests (and I also tested this manually, just to be sure).

See #349