PHPUnit factory: Don't extend class registry
We can’t extend the class registry because WordPoints won’t be loaded
during the uninstall tests.

We actually don’t use the factory in the uninstall tests, so we could
just skip loading it then I guess. But it is better this way, to have
the tests cleanly separate from WordPoints core code.

See #321