Move uninstall tests to separate test suite
When the test is in the same suite as the other tests, all of the
PHPUnit tests are loaded when it is run. This is a problem because as
part of that, all of the data providers will be called, and some of
these use core functions and classes, which won’t be loaded when the
tests are being run. Putting the uninstall test in its own suite
prevents this from happening. However, it also means that they the test
has to be run with `—testsuite uninstall —group uninstall`. The test
suite part is needed to select that test suite, and the group part is
needed because we still need to exclude the uninstall group. The latter
is because PHPUnit will run all test suites when called from the
command line without a particular test suite provided, which we need to
maintain for backward compatibility. So we need to continue to keep the
uninstall group designation for this test, and exclude that group by
default. We could have moved the suite to a completely separate XML
config file instead, but I didn’t want to do that at this time because
of how the dev-lib detects uninstall tests, etc.

See #321