Don't show notice when network-installing ranks
I’ve deprecated the `do_per_site_install()` boolean method in favor of
a method that returns a bit mask value, giving us greater granularity
in how we respond to the returned value. It is therefore possible to
now skip the per-site installation when network installing without
indicating that the user needs to perform some parts of the install
process manually. We employ this new feature within the ranks component
to ensure that we don’t show users a notice about a part of the install
process that we just don’t need at all for this component.

Along the way, I have decided to make the check for whether manually
install is required completely independent from the check for whether
the install is skipped. This would permit a module/component to do a
partial per-site install, while skipping part of it, if they so chose.

I’ve also updated the descriptions in the docblocks of the
`set_network_install_skipped()` and `set_network_update_skipped()`
methods, since they don’t mean that the network portion of the
uninstall was skipped, but rather that the per-site part of the network
install was skipped.

Unit tests are included for basic installation as well as the behavior
in question.

Fixes #360