Don't run per-site updates if there are none
Previously we were looping through all sites when updating a
network-activated entity on multisite even when there were no per-site
updates that needed to be performed.

Also, we were informing the user that some parts of the update process
were skipped and needed to be performed manually, even when there were
no per-site updates. This is a similar issue to #360.

I’ve fixed both of these and included basic unit tests for the
`update()` method.

Fixes #392